# NS-115: Implement efficient root hierarchy queries in LanceDB data store