--- a/Cargo.toml
+++ b/Cargo.toml
@@ -39,3 +39,8 @@ default = []
 
 [dev-dependencies]
 tokio-test = "0.4"
+criterion = { version = "0.5", features = ["html_reports"] }
+
+[[bench]]
+name = "hierarchy_performance"
+harness = false