--- a/tests/ns115_root_hierarchy_tests.rs
+++ b/tests/ns115_root_hierarchy_tests.rs
@@ -5,7 +5,7 @@ use nodespace_data_store::{DataStore, LanceDataStore};
 #[tokio::test]
 async fn test_root_hierarchy_optimization() -> NodeSpaceResult<()> {
     let data_store = LanceDataStore::new("data/test_ns115_root_hierarchy.db").await?;
-    
+
     // Create a hierarchical structure
     let root_id = NodeId::from_string("project-alpha".to_string());
     let root_node = Node {
@@ -22,9 +22,9 @@ async fn test_root_hierarchy_optimization() -> NodeSpaceResult<()> {
         next_sibling: None,
         previous_sibling: None,
     };
-    
+
     data_store.store_node(root_node).await?;
-    
+
     // Create child nodes
     let mut child_ids = Vec::new();
     for i in 0..5 {
@@ -44,11 +44,11 @@ async fn test_root_hierarchy_optimization() -> NodeSpaceResult<()> {
             next_sibling: None,
             previous_sibling: None,
         };
-        
+
         data_store.store_node(child_node).await?;
         child_ids.push(child_id);
     }
-    
+
     // Create some unrelated nodes (different root)
     let other_root_id = NodeId::from_string("project-beta".to_string());
     let other_root_node = Node {
@@ -65,13 +65,13 @@ async fn test_root_hierarchy_optimization() -> NodeSpaceResult<()> {
         next_sibling: None,
         previous_sibling: None,
     };
-    
+
     data_store.store_node(other_root_node).await?;
-    
+
     // Test 1: get_nodes_by_root should return only Project Alpha nodes
     let alpha_nodes = data_store.get_nodes_by_root(&root_id).await?;
     assert_eq!(alpha_nodes.len(), 6); // 1 root + 5 children
-    
+
     // Verify all returned nodes have the correct root_id
     for node in &alpha_nodes {
         if let Some(metadata) = &node.metadata {
@@ -80,11 +80,13 @@ async fn test_root_hierarchy_optimization() -> NodeSpaceResult<()> {
             }
         }
     }
-    
+
     // Test 2: get_nodes_by_root_and_type should filter by type
-    let alpha_tasks = data_store.get_nodes_by_root_and_type(&root_id, "task").await?;
+    let alpha_tasks = data_store
+        .get_nodes_by_root_and_type(&root_id, "task")
+        .await?;
     assert_eq!(alpha_tasks.len(), 5); // Only the 5 task nodes
-    
+
     // Verify all returned nodes are tasks
     for node in &alpha_tasks {
         if let Some(metadata) = &node.metadata {
@@ -93,12 +95,12 @@ async fn test_root_hierarchy_optimization() -> NodeSpaceResult<()> {
             }
         }
     }
-    
+
     // Test 3: Verify performance - O(1) vs O(N) comparison
     let start_time = std::time::Instant::now();
     let _result1 = data_store.get_nodes_by_root(&root_id).await?;
     let optimized_duration = start_time.elapsed();
-    
+
     let start_time = std::time::Instant::now();
     let all_nodes = data_store.query_nodes("").await?;
     let _filtered_nodes: Vec<_> = all_nodes
@@ -112,15 +114,21 @@ async fn test_root_hierarchy_optimization() -> NodeSpaceResult<()> {
         })
         .collect();
     let unoptimized_duration = start_time.elapsed();
-    
-    println!("🚀 Optimized query (get_nodes_by_root): {:?}", optimized_duration);
-    println!("🐌 Unoptimized query (query_nodes + filter): {:?}", unoptimized_duration);
-    
+
+    println!(
+        "🚀 Optimized query (get_nodes_by_root): {:?}",
+        optimized_duration
+    );
+    println!(
+        "🐌 Unoptimized query (query_nodes + filter): {:?}",
+        unoptimized_duration
+    );
+
     // Test 4: Empty result for non-existent root
     let empty_root_id = NodeId::from_string("non-existent".to_string());
     let empty_result = data_store.get_nodes_by_root(&empty_root_id).await?;
     assert_eq!(empty_result.len(), 0);
-    
+
     println!("✅ NS-115 root hierarchy optimization tests passed!");
     Ok(())
 }
@@ -128,7 +136,7 @@ async fn test_root_hierarchy_optimization() -> NodeSpaceResult<()> {
 #[tokio::test]
 async fn test_root_id_schema_fields() -> NodeSpaceResult<()> {
     let data_store = LanceDataStore::new("data/test_ns115_schema.db").await?;
-    
+
     // Test that root_id fields are properly stored and retrieved
     let test_id = NodeId::from_string("schema-test".to_string());
     let test_node = Node {
@@ -145,13 +153,13 @@ async fn test_root_id_schema_fields() -> NodeSpaceResult<()> {
         next_sibling: None,
         previous_sibling: None,
     };
-    
+
     data_store.store_node(test_node).await?;
-    
+
     // Retrieve and verify
     let retrieved_node = data_store.get_node(&test_id).await?;
     assert!(retrieved_node.is_some());
-    
+
     let node = retrieved_node.unwrap();
     if let Some(metadata) = &node.metadata {
         assert_eq!(
@@ -163,7 +171,7 @@ async fn test_root_id_schema_fields() -> NodeSpaceResult<()> {
             Some("test_hierarchy")
         );
     }
-    
+
     println!("✅ NS-115 schema fields validation passed!");
     Ok(())
 }